feat: bring grant response type into shared package (#1377)
- Bring grant response type into shared package
