+++ packages/wallet/shared/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './account'
 export * from './asset'
 export * from './rate'
+export * from './grant'
